requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
chess==1.9.4
cmake==3.26.4
customtkinter==5.1.3
darkdetect==0.8.0
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
